Add new config for non-managed models

+++ dynamic_models/models.py
@@ -11,25 +11,36 @@ from dynamic_models.utils import ModelRegistry
 
 
 class ModelSchema(models.Model):
-    name = models.CharField(max_length=32, unique=True)
+    name = models.CharField(max_length=250, unique=True)
     db_name = models.CharField(max_length=32, default=DEFAULT_DB_ALIAS)
+    managed = models.BooleanField(default=True)
+    db_table_name = models.CharField(null=True, max_length=250)
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self._registry = ModelRegistry(self.app_label)
         self._initial_name = self.name
         initial_model = self.get_registered_model()
-        self._schema_editor = ModelSchemaEditor(initial_model, db_name=self.db_name)
+        self._schema_editor = (
+            ModelSchemaEditor(
+                initial_model=initial_model,
+                db_name=self.db_name
+            )
+            if self.managed
+            else None
+        )
 
     def save(self, **kwargs):
         super().save(**kwargs)
         cache.update_last_modified(self.model_name)
         cache.update_last_modified(self.initial_model_name)
-        self._schema_editor.update_table(self._factory.make_model())
+        if self._schema_editor:
+            self._schema_editor.update_table(self._factory.make_model())
         self._initial_name = self.name
 
     def delete(self, **kwargs):
-        self._schema_editor.drop_table(self.as_model())
+        if self._schema_editor:
+            self._schema_editor.drop_table(self.as_model())
         self._factory.destroy_model()
         cache.clear_last_modified(self.initial_model_name)
         super().delete(**kwargs)
@@ -59,8 +70,11 @@ class ModelSchema(models.Model):
 
     @property
     def db_table(self):
-        parts = (self.app_label, slugify(self.name).replace("-", "_"))
-        return "_".join(parts)
+        return self.db_table_name if self.db_table_name else self._default_db_table_name() 
+    
+    def _default_db_table_name(self):
+        safe_name = slugify(self.name).replace("-", "_")
+        return f"{self.app_label}_{safe_name}"
 
     def as_model(self):
         return self._factory.get_model()
@@ -126,8 +140,13 @@ class FieldSchema(models.Model):
         self._initial_name = self.name
         self._initial_null = self.null
         self._initial_field = self.get_registered_model_field()
-        self._schema_editor = FieldSchemaEditor(
-            self._initial_field, db_name=self.model_schema.db_name
+        self._schema_editor = (
+            FieldSchemaEditor(
+                initial_field=self._initial_field,
+                db_name=self.model_schema.db_name
+            )
+            if self.model_schema.managed
+            else None
         )
 
     def save(self, **kwargs):
@@ -135,11 +154,13 @@ class FieldSchema(models.Model):
         super().save(**kwargs)
         self.update_last_modified()
         model, field = self._get_model_with_field()
-        self._schema_editor.update_column(model, field)
+        if self._schema_editor:
+            self._schema_editor.update_column(model, field)
 
     def delete(self, **kwargs):
         model, field = self._get_model_with_field()
-        self._schema_editor.drop_column(model, field)
+        if self._schema_editor:
+            self._schema_editor.drop_column(model, field)
         self.update_last_modified()
         super().delete(**kwargs)
